Move `split_at_mut_unchecked` to utils crate; remove unused import.
