+++ util/src/lib.rs
@@ -168,3 +168,27 @@ pub fn pretty_name<T>() -> String {
     }
     result
 }
+
+/// Copied from Rust nightly sources
+#[inline(always)]
+unsafe fn from_raw_parts_mut<'a, T>(data: *mut T, len: usize) -> &'a mut [T] {
+    unsafe { &mut *core::ptr::slice_from_raw_parts_mut(data, len) }
+}
+
+/// Copied from Rust nightly sources
+#[inline(always)]
+pub unsafe fn split_at_mut_unchecked<T>(v: &mut [T], mid: usize) -> (&mut [T], &mut [T]) {
+    let len = v.len();
+    let ptr = v.as_mut_ptr();
+
+    // SAFETY: Caller has to check that `0 <= mid <= self.len()`.
+    //
+    // `[ptr; mid]` and `[mid; len]` are not overlapping, so returning
+    // a mutable reference is fine.
+    unsafe {
+        (
+            from_raw_parts_mut(ptr, mid),
+            from_raw_parts_mut(ptr.add(mid), len - mid),
+        )
+    }
+}
